✨ feat: add support for a placeholder element in the useDelayedRender hook
Had to increase the delay on the Wallet dropdown rendering as sometimes the polkadot{.js} extension
would take a bit of time to become available. But, I didn't like how the Wallet button would "pop"
in late. It was pretty jarring, like an old school FUOC.

By adding an optional placeholder to useDelayedRender we can render a "disabled" button until the
longer timeout has passed and the actual Wallet dropdown button can be rendered.